Pole Price urzadzen jest teraz typu float pozwalajac na ceny z ulamkiem

+++ ProjektAPBDCw2/Camera.cs
@@ -5,7 +5,7 @@ public class Camera : Device
     public int Resolution { get; }
     public bool LensExchangeable { get; }
     
-    public Camera(string name, Manufacturer manufacturer, int price, int resolution, bool lensExchangeable) : base(name, manufacturer, price)
+    public Camera(string name, Manufacturer manufacturer, float price, int resolution, bool lensExchangeable) : base(name, manufacturer, price)
     {
         Resolution = resolution;
         LensExchangeable = lensExchangeable;
