+++ ProjektAPBDCw2/Laptop.cs
@@ -5,7 +5,7 @@ public class Laptop : Device
     public float BatteryTime { get; }
     public bool HasWebcam { get; }
 
-    public Laptop(string name, Manufacturer manufacturer, int price, float batteryTime, bool hasWebcam) : base(name, manufacturer, price)
+    public Laptop(string name, Manufacturer manufacturer, float price, float batteryTime, bool hasWebcam) : base(name, manufacturer, price)
     {
         BatteryTime = batteryTime;
         HasWebcam = hasWebcam;
